added sortOffers
