+++ src/components/pages/OffererPage.js
@@ -11,27 +11,12 @@ import { requestData } from '../../reducers/data'
 import { setUserOfferer } from '../../reducers/user'
 
 class OffererPage extends Component {
-  handleSetUserOfferer = props => {
-    const { offererId, setUserOfferer, user } = props
-    user && user.userOfferers && setUserOfferer(offererId)
-  }
+
   componentWillMount() {
     const { requestData } = this.props
     requestData('GET', 'providers')
-    this.handleSetUserOfferer(this.props)
   }
-  componentWillReceiveProps(nextProps) {
-    const { offererId, requestData, user } = nextProps
-    if (
-      user &&
-      user.userOfferers !== (this.props.user && this.props.user.userOfferers)
-    ) {
-      this.handleSetUserOfferer(nextProps)
-    }
-    if (user && this.props.user) {
-      requestData('GET', `offerers/${offererId}`, { key: 'offerers' })
-    }
-  }
+
   render() {
     return (
       <main className="page offerer-page p2">
@@ -48,5 +33,8 @@ class OffererPage extends Component {
 
 export default compose(
   withLogin({ isRequired: true }),
-  connect(state => ({ user: state.user }), { requestData, setUserOfferer })
+  connect(
+    state => ({ user: state.user }),
+    { requestData, setUserOfferer }
+  )
 )(OffererPage)
